requirements.txt
-e git+https://github.com/vyscond/nest.git@7d4f90d6530dc17b43cd0ad2899aa3692c1e3958#egg=nest-master
pyaml==15.6.3
pyflakes==0.9.2
PyYAML==3.11
wheel==0.24.0
